requirements.txt
pika==1.3.2
prefect==2.14.21
feedparser==6.0.11
kafka-python==2.0.2
